chore: move binding utils
