+++ crates/node_binding/src/lib.rs
@@ -11,7 +11,6 @@ use std::collections::HashSet;
 use std::pin::Pin;
 use std::sync::atomic::{AtomicU32, Ordering};
 
-use dashmap::DashMap;
 use napi::bindgen_prelude::*;
 use once_cell::sync::Lazy;
 use rspack_core::PluginExt;
@@ -41,91 +40,11 @@ static GLOBAL: mimalloc_rust::GlobalMiMalloc = mimalloc_rust::GlobalMiMalloc;
 #[global_allocator]
 static GLOBAL: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
 
-// **Note** that Node's main thread and the worker thread share the same binding context. Using `Mutex<HashMap>` would cause deadlocks if multiple compilers exist.
-struct SingleThreadedHashMap<K, V>(DashMap<K, V>);
-
-impl<K, V> SingleThreadedHashMap<K, V>
-where
-  K: Eq + std::hash::Hash + std::fmt::Display,
-{
-  /// Acquire a mutable reference to the inner hashmap.
-  ///
-  /// Safety: Mutable reference can almost let you do anything you want, this is intended to be used from the thread where the map was created.
-  #[allow(unused)]
-  unsafe fn borrow_mut<F, R>(&self, key: &K, f: F) -> Result<R>
-  where
-    F: FnOnce(&mut V) -> Result<R>,
-  {
-    let mut inner = self.0.get_mut(key).ok_or_else(|| {
-      napi::Error::from_reason(format!(
-        "Failed to find key {key} for single-threaded hashmap",
-      ))
-    })?;
-
-    f(&mut inner)
-  }
-
-  /// Acquire a shared reference to the inner hashmap.
-  ///
-  /// Safety: It's not thread-safe if a value is not safe to modify cross thread boundary, so this is intended to be used from the thread where the map was created.
-  #[allow(unused)]
-  unsafe fn borrow<F, R>(&self, key: &K, f: F) -> Result<R>
-  where
-    F: FnOnce(&V) -> Result<R>,
-  {
-    let inner = self.0.get(key).ok_or_else(|| {
-      napi::Error::from_reason(format!(
-        "Failed to find key {key} for single-threaded hashmap",
-      ))
-    })?;
-
-    f(&*inner)
-  }
-
-  /// Insert a value into the map.
-  ///
-  /// Safety: It's not thread-safe if a value has thread affinity, so this is intended to be used from the thread where the map was created.
-  #[allow(unused)]
-  unsafe fn insert_if_vacant(&self, key: K, value: V) -> Result<()> {
-    if let dashmap::mapref::entry::Entry::Vacant(vacant) = self.0.entry(key) {
-      vacant.insert(value);
-      Ok(())
-    } else {
-      Err(napi::Error::from_reason(
-        "Failed to insert on single-threaded hashmap as it's not vacant",
-      ))
-    }
-  }
-
-  /// Remove a value from the map.
-  ///
-  /// See: [DashMap::remove] for more details. https://docs.rs/dashmap/latest/dashmap/struct.DashMap.html#method.remove
-  ///
-  /// Safety: It's not thread-safe if a value has thread affinity, so this is intended to be used from the thread where the map was created.
-  #[allow(unused)]
-  unsafe fn remove(&self, key: &K) -> Option<V> {
-    self.0.remove(key).map(|(_, v)| v)
-  }
-}
-
-impl<K, V> Default for SingleThreadedHashMap<K, V>
-where
-  K: Eq + std::hash::Hash,
-{
-  fn default() -> Self {
-    Self(Default::default())
-  }
-}
-
-// Safety: Methods are already marked as unsafe.
-unsafe impl<K, V> Send for SingleThreadedHashMap<K, V> {}
-unsafe impl<K, V> Sync for SingleThreadedHashMap<K, V> {}
-
 static COMPILERS: Lazy<
   SingleThreadedHashMap<CompilerId, Pin<Box<rspack_core::Compiler<AsyncNodeWritableFileSystem>>>>,
 > = Lazy::new(Default::default);
 
-static COMPILER_ID: AtomicU32 = AtomicU32::new(1);
+static NEXT_COMPILER_ID: AtomicU32 = AtomicU32::new(0);
 
 type CompilerId = u32;
 
@@ -177,7 +96,7 @@ impl Rspack {
         .map_err(|e| Error::from_reason(format!("Failed to create writable filesystem: {e}",)))?,
     );
 
-    let id = COMPILER_ID.fetch_add(1, Ordering::SeqCst);
+    let id = NEXT_COMPILER_ID.fetch_add(1, Ordering::SeqCst);
     unsafe { COMPILERS.insert_if_vacant(id, Box::pin(rspack)) }?;
 
     Ok(Self { id, disabled_hooks })
