+++ crates/node_binding/src/utils.rs
@@ -3,6 +3,7 @@ use std::ffi::CStr;
 use std::io::Write;
 use std::ptr;
 
+use dashmap::DashMap;
 use futures::Future;
 use napi::bindgen_prelude::*;
 use napi::{check_status, Env, Error, JsFunction, JsUnknown, NapiRaw, Result};
@@ -181,3 +182,81 @@ where
 
   Ok(())
 }
+
+// **Note** that Node's main thread and the worker thread share the same binding context. Using `Mutex<HashMap>` would cause deadlocks if multiple compilers exist.
+pub(crate) struct SingleThreadedHashMap<K, V>(DashMap<K, V>);
+
+impl<K, V> SingleThreadedHashMap<K, V>
+where
+  K: Eq + std::hash::Hash + std::fmt::Display,
+{
+  /// Acquire a mutable reference to the inner hashmap.
+  ///
+  /// Safety: Mutable reference can almost let you do anything you want, this is intended to be used from the thread where the map was created.
+  pub(crate) unsafe fn borrow_mut<F, R>(&self, key: &K, f: F) -> Result<R>
+  where
+    F: FnOnce(&mut V) -> Result<R>,
+  {
+    let mut inner = self.0.get_mut(key).ok_or_else(|| {
+      napi::Error::from_reason(format!(
+        "Failed to find key {key} for single-threaded hashmap",
+      ))
+    })?;
+
+    f(&mut inner)
+  }
+
+  /// Acquire a shared reference to the inner hashmap.
+  ///
+  /// Safety: It's not thread-safe if a value is not safe to modify cross thread boundary, so this is intended to be used from the thread where the map was created.
+  #[allow(unused)]
+  pub(crate) unsafe fn borrow<F, R>(&self, key: &K, f: F) -> Result<R>
+  where
+    F: FnOnce(&V) -> Result<R>,
+  {
+    let inner = self.0.get(key).ok_or_else(|| {
+      napi::Error::from_reason(format!(
+        "Failed to find key {key} for single-threaded hashmap",
+      ))
+    })?;
+
+    f(&*inner)
+  }
+
+  /// Insert a value into the map.
+  ///
+  /// Safety: It's not thread-safe if a value has thread affinity, so this is intended to be used from the thread where the map was created.
+  pub(crate) unsafe fn insert_if_vacant(&self, key: K, value: V) -> Result<()> {
+    if let dashmap::mapref::entry::Entry::Vacant(vacant) = self.0.entry(key) {
+      vacant.insert(value);
+      Ok(())
+    } else {
+      Err(napi::Error::from_reason(
+        "Failed to insert on single-threaded hashmap as it's not vacant",
+      ))
+    }
+  }
+
+  /// Remove a value from the map.
+  ///
+  /// See: [DashMap::remove] for more details. https://docs.rs/dashmap/latest/dashmap/struct.DashMap.html#method.remove
+  ///
+  /// Safety: It's not thread-safe if a value has thread affinity, so this is intended to be used from the thread where the map was created.
+  #[allow(unused)]
+  pub(crate) unsafe fn remove(&self, key: &K) -> Option<V> {
+    self.0.remove(key).map(|(_, v)| v)
+  }
+}
+
+impl<K, V> Default for SingleThreadedHashMap<K, V>
+where
+  K: Eq + std::hash::Hash,
+{
+  fn default() -> Self {
+    Self(Default::default())
+  }
+}
+
+// Safety: Methods are already marked as unsafe.
+unsafe impl<K, V> Send for SingleThreadedHashMap<K, V> {}
+unsafe impl<K, V> Sync for SingleThreadedHashMap<K, V> {}
